04_LoRA_FineTuning_Project/requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets
peft
accelerate
scikit-learn
python-dotenv
bitsandbytes
tensorboard
tqdm
streamlit
uvicorn
fastapi